fix (security): Bad handling of StartTLS exposes credentials
When using the `start-tls` option, the `session.StartTLS()` call was
happening after the call to `session.Bind()`, which caused the LDAP
credentials to be exposed in plaintext. This change modifies the order
so that the TLS connection is established before the bind, the way it
should have been done in the first place.

This change also switches from printing the search results using the
structured Logger interface to just printing the JSON string.

## Changed

- Move the StartTLS call before the bind so credentials are not exposed
- Print search results using `fmt.Println` instead of using the Logger